Add Authors.tsx component for rendering work / manifestation creators
